+++ PhotocentreDatabase2/src/main/resources/db/migration/V6__Create_Orders.sql
@@ -8,7 +8,7 @@ create table if not exists Orders
   order_date            date default current_date                                                                                                   not null,
   order_completion_date date,
   order_type            varchar(256)                                                                                                                not null,
-  branch_office_id      int references branch_offices (branch_office_id) on delete cascade on update cascade                                        not null,
-  kiosk_id              int references kiosks (kiosk_id) on delete cascade on update cascade,
-  customer_id           int references customers (customer_id) on delete cascade on update cascade                                                  not null
+  branch_office_id      bigint references branch_offices (branch_office_id) on delete cascade on update cascade                                        not null,
+  kiosk_id              bigint references kiosks (kiosk_id) on delete cascade on update cascade,
+  customer_id           bigint references customers (customer_id) on delete cascade on update cascade                                                  not null
 );
